+++ R/IncMat_QTL_MAF.R
@@ -81,8 +81,12 @@ IncMat_QTL_MAF <- function(QTL, Q.eff_i, Q.pos_i, mppData, ref.par = NULL){
         
       } else {
         
-        Q.mat <- cbind(Q.mat, QTL_i[, names(sort(all.freq))])
-        allele_ord_i <- rev(colnames(Q.mat))
+        # Q.mat <- cbind(Q.mat, QTL_i[, names(sort(all.freq))])
+        # allele_ord_i <- rev(colnames(Q.mat))
+        
+        name.order <- names(sort(all.freq))
+        Q.mat <- cbind(Q.mat, QTL_i[, name.order])
+        allele_ord_i <- rev(name.order)
         
       }