+++ R/color.code.R
@@ -5,8 +5,54 @@
 # color indicator function
 
 # all p-values from a nested model are transformed in a value between
-# -5 and 5 which will give a indication of colour (blue or red) and the
+# -6 and 6 which will give a indication of colour (blue or red) and the
 # intensity of the colour
+
+# color.code <- function(x){
+#   
+#   # red colours
+#   
+#   if(x<0){
+#     
+#     x <- abs(x)
+#     
+#     if(x>=0.05){
+#       col <- 0
+#     }else if((0.05>x) & (1e-2<=x)){
+#       col <- -1
+#     }else if((1e-2>x) & (1e-3<=x)){
+#       col <- -2
+#     }else if((1e-3>x) & (1e-4<=x)){
+#       col <- -3
+#     }else if((1e-4>x) & (1e-5<=x)){
+#       col <- -4
+#     }else if(1e-5>x){
+#       col <- -5
+#     }
+#     
+#   }else{
+#     
+#     # blue colors
+#     
+#     if(x>=0.05){
+#       col <- 0
+#     }else if((0.05>x) & (1e-2<=x)){
+#       col <- 1
+#     }else if((1e-2>x) & (1e-3<=x)){
+#       col <- 2
+#     }else if((1e-3>x) & (1e-4<=x)){
+#       col <- 3
+#     }else if((1e-4>x) & (1e-5<=x)){
+#       col <- 4
+#     }else if(1e-5>x){
+#       col <- 5
+#     }
+#     
+#   }
+#   
+#   return(col)
+#   
+# }
 
 color.code <- function(x){
   
@@ -18,16 +64,10 @@ color.code <- function(x){
     
     if(x>=0.05){
       col <- 0
-    }else if((0.05>x) & (1e-2<=x)){
-      col <- -1
-    }else if((1e-2>x) & (1e-3<=x)){
-      col <- -2
-    }else if((1e-3>x) & (1e-4<=x)){
-      col <- -3
-    }else if((1e-4>x) & (1e-5<=x)){
-      col <- -4
-    }else if(1e-5>x){
-      col <- -5
+    }else if((0.05>x) & (1e-6<=x)){
+      col <- -(-log10(x))
+    }else if(1e-6>x){
+      col <- -6
     }
     
   }else{
@@ -36,16 +76,10 @@ color.code <- function(x){
     
     if(x>=0.05){
       col <- 0
-    }else if((0.05>x) & (1e-2<=x)){
-      col <- 1
-    }else if((1e-2>x) & (1e-3<=x)){
-      col <- 2
-    }else if((1e-3>x) & (1e-4<=x)){
-      col <- 3
-    }else if((1e-4>x) & (1e-5<=x)){
-      col <- 4
-    }else if(1e-5>x){
-      col <- 5
+    }else if((0.05>x) & (1e-6<=x)){
+      col <- (-log10(x))
+    }else if(1e-6>x){
+      col <- 6
     }
     
   }